Add GraphHealthDetails and GraphHealthStatus models; update stats router imports
